Time: 0 ms (100.00%) | Memory: 11.3 MB (94.55%) - LeetSync

+++ 3227-find-missing-and-repeated-values/find-missing-and-repeated-values.c
@@ -13,9 +13,9 @@ int* findMissingAndRepeatedValues(int** grid, int gridSize, int* gridColSize, in
         // Calculate expected sum of numbers and squared numbers in grid
         // sum of n natural numbers = [n(n+1)]/2
         // sum of n squared natural numbers = [n(n+1)(2n+1)]/6
-        long long n = gridSize * gridSize;
+        int n = gridSize * gridSize;
         int perfect_sum = n * (n + 1) / 2;
-        long long perfect_sqr_sum = n * (n + 1) * (2 * n + 1) / 6;
+        long long perfect_sqr_sum = (long long)n * (n + 1) * (2 * (long long)n + 1) / 6;
         
         // Breaking down expressions
         // ==================================================
